print_code: allow user to pass in caller line or path and line number
